backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.40
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
cryptography==44.0.2
alembic==1.15.2
